% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.r
\name{createBenchmarkWts}
\alias{createBenchmarkWts}
\title{Create Benchmark Weights}
\usage{
createBenchmarkWts(benchmark.txt, investor)
}
\arguments{
\item{benchmark.txt}{string of benchmark definition}

\item{investor}{investor class object containing initialValues$PreTaxPct to allocate Bonds appropriately.}
}
\value{
a list with the two vectors and a the string defining the benchmark.
}
\description{
Given a benchmark definition in a string assignes the weights to each class.
It creates a vector for the base classes (n asset classes) and a vector with the tax status (3n).
If the benchmark definition has a weight to an asset class called Bonds, it will assign a percentage of that weight
to the taxable account in Munis and a percentage into USCoreBonds for the deferred and exempt accounts.
}
