% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.r
\name{readACNameFile}
\alias{readACNameFile}
\title{Read Asset Class Name File}
\usage{
readACNameFile(fn, sheet = 1)
}
\arguments{
\item{fn}{name of file}

\item{sheet}{which sheet to read, defaults to first}
}
\value{

}
\description{
This file contains both a mapping of RT asset class names to RAFI and characteristics of the asset classes
}
