% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.r
\name{readExcelFile}
\alias{readExcelFile}
\title{Reads and Excel file}
\usage{
readExcelFile(fileName, sheet = NULL)
}
\arguments{
\item{fileName}{name of Excel file}

\item{sheet}{sheet to read, default, NULL, leads to first}
}
\value{

}
\description{
Makes sure the names are legit in R
}
