% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.r
\name{readPortfolioHoldings}
\alias{readPortfolioHoldings}
\title{Read Portfolio Holdings from File}
\usage{
readPortfolioHoldings(portfolioName = NULL, portfolioHoldingsFN)
}
\arguments{
\item{portfolioName}{name of portfolio for which to filter holdings. If NULL all holdings for all portfolios are returned.}

\item{portfolioHoldingsFN}{name of Excel file with the holdings}
}
\value{

}
\description{
# Read in a file exported from Black Diamond / Data Mining / Rebalance Template extract a portfolio
 Munge a bit
}
