% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.r
\name{readRAFIFile}
\alias{readRAFIFile}
\title{Read RAFI XLSX file}
\usage{
readRAFIFile(fn, rafi_xlranges = NULL)
}
\arguments{
\item{fn}{name of XLSX file including path}

\item{rafi_xlranges}{list with 4 elements: return, corr, cov, and date are the Excel ranges
which contain the respective data.}
}
\value{
A list with 4 items: ret is the return data, corr is the
  correlation data, cov is the covariance matrix and as_of_date is the date of the file
}
\description{
This function reads a set of returns and correlations from XLS file
downloaded from the RAFI website.  No processing is done. This function
allows one to see the data in an unprocessed form.
}
\keyword{allocation}
\keyword{asset}
\keyword{efficient}
\keyword{frontier}
